refactor: gameSlot, gameSlotDetail 합치기
